# go/MongoWebProject/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server",
        "title": "4.Mongo Web Project",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/products": {
            "get": {
                "description": "Returns the list of the products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get all products",
                "responses": {}
            },
            "post": {
                "description": "Returns the product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Save product",
                "parameters": [
                    {
                        "description": "shop product",
                        "name": "Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ProductForm"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Returns the product by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product by id"
                ],
                "summary": "Get product by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "request.ProductForm": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}